pyproject.toml
[project]
name = "article-to-speech"
version = "0.1.0"
description = "Dockerized Telegram article-to-audio automation using the ChatGPT website."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "edge-tts>=7.2.8",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "playwright>=1.53.0",
    "python-dateutil>=2.9.0.post0",
    "readability-lxml>=0.8.1",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
]

[project.scripts]
article-to-speech = "article_to_speech.cli:main"

[dependency-groups]
dev = [
    "click>=8.3.1",
    "pyright[nodejs]>=1.1.408",
    "pysonar>=1.3.0.4086",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
    "truststore>=0.10.4",
    "vulture>=2.14",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/article_to_speech"]

[tool.pytest.ini_options]
addopts = "-q -s"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]

[tool.vulture]
paths = ["src"]
min_confidence = 80
